remsim_client: add support to set sim presence pin polarity
The sim presence pin allows the simtrace2 to inform the modem
about the presence of the simcard. On a generic simcard slot a
button is pressed by the simcard which generate a high or low voltage level.
Even there are specifications of minipcie or ngff defining this signal, certain
modems don't follow the specification and require a different signal level.

Related: SYS#6946
Depends-on: Iacde63e667472a85a5f18cb4ca99dcfa1a84dd59 (simtrace2)
Change-Id: Iaa61967d2e8cd5c2269b73f4e65b42072e0ec513